requirements.txt
Flask~=3.1.0
flask-cors~=5.0.0
google-generativeai
chromadb~=0.6.3
langchain-community
sentence-transformers
protobuf~=5.29.3
langchain~=0.3.17
python-dotenv
pymongo
certifi